# server/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Backend API for the SITCON Camp 2026 game.",
        "title": "Camp 2026 Game API",
        "contact": {},
        "version": "0.1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/me/status": {
            "get": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Returns the authenticated player's profile summary, team, and open power total.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "me"
                ],
                "summary": "Get current player status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/me.StatusResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/qr/resolve": {
            "post": {
                "security": [
                    {
                        "AuthCookieAuth": []
                    }
                ],
                "description": "Staff-only endpoint. Resolves a player QR code identifier into a player summary without exposing auth credentials.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "qr"
                ],
                "summary": "Resolve player QR code",
                "parameters": [
                    {
                        "description": "QR resolve request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/qr.ResolveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/qr.ResolveResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "422": {
                        "description": "Unprocessable Entity",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/httpx.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "httpx.ErrorDetail": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "httpx.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/httpx.ErrorDetail"
                    }
                },
                "instance": {
                    "type": "string"
                },
                "requestId": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "me.StatusResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string",
                    "example": "https://example.test/avatar/alice.png"
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "openPower": {
                    "type": "integer",
                    "example": 1280
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "role": {
                    "type": "string",
                    "example": "staff"
                },
                "team": {
                    "$ref": "#/definitions/me.TeamResponse"
                },
                "teamMembers": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/me.TeamMemberResponse"
                    }
                }
            }
        },
        "me.TeamMemberResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string",
                    "example": "https://example.test/avatar/alice.png"
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "role": {
                    "type": "string",
                    "example": "staff"
                }
            }
        },
...
        "me.TeamResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Blue Team"
                },
                "teamId": {
                    "type": "string",
                    "example": "8M4RXP"
                }
            }
        },
        "qr.PlayerSummaryResponse": {
            "type": "object",
            "properties": {
                "avatarUrl": {
                    "type": "string",
                    "example": "https://example.test/avatar/alice.png"
                },
                "nickname": {
                    "type": "string",
                    "example": "Alice"
                },
                "openPower": {
                    "type": "integer",
                    "example": 1280
                },
                "playerId": {
                    "type": "string",
                    "example": "7H9K2Q"
                },
                "team": {
                    "$ref": "#/definitions/qr.TeamResponse"
                }
            }
        },
        "qr.ResolveRequest": {
            "type": "object",
            "required": [
                "qrcodeToken"
            ],
            "properties": {
                "qrcodeToken": {
                    "type": "string",
                    "maxLength": 512,
                    "minLength": 4,
                    "example": "qr_6H_x7lM20CK8BBnPfwEG1Ei97-PM9ZGr8Dy9yW-BYok"
                }
            }
        },
        "qr.ResolveResponse": {
            "type": "object",
            "properties": {
                "player": {
                    "$ref": "#/definitions/qr.PlayerSummaryResponse"
                }
            }
        },
        "qr.TeamResponse": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string",
                    "example": "Blue Team"
                },
                "teamId": {
                    "type": "string",
                    "example": "8M4RXP"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "AuthCookieAuth": {
            "description": "User auth cookie. Send as `Cookie: camp2026_auth=\u003cbase64url-auth-token\u003e`.",
            "type": "apiKey",
            "name": "Cookie",
            "in": "header"
        }
    }
}